Base_Implementation/JAVA Implementation/src/entities/OpNode.java
package entities;

public class OpNode extends ExprNode {
    private final ExprNode left;
    private final ExprNode right;
    private final String operator;

    public OpNode(ExprNode left, ExprNode right, String operator) {
        this.left = left;
        this.right = right;
        this.operator = operator;
    }

    public ExprNode getLeft() {
        return left;
    }

    public ExprNode getRight() {
        return right;
    }

    public String getOperator() {
        return operator;
    }

    @Override
    public String toString() {
        return "(" + left.toString() + " " + operator + " " + right.toString() + ")";
    }
}
